fix test

+++ plugin/dapp/cross2eth/ebrelayer/utils/utils_test.go
@@ -8,10 +8,10 @@ import (
 )
 
 func Test_decimals(t *testing.T) {
-	value2comp := 10
+	value2comp := int64(10)
 	for i := 1; i <= 10; i++ {
 		value, ok := Decimal2value[i]
-		require.Equal(t, value, int64(value2comp))
+		require.Equal(t, value, value2comp)
 		require.Equal(t, ok, true)
 		fmt.Println("value=", value)
 		value2comp = value2comp * 10
